requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
jinja2
sqlalchemy